+++ analyzers/__init__.py
@@ -1,3 +1,4 @@
+from analyzer import Analyzer
 from error_stats import AnalyzeErrors
 from coverage_modes import AnalyzeCoverageModes
 from coverage_stop_areas import AnalyzeCoverageStopAreas
@@ -6,4 +7,3 @@ from requests_calls import AnalyzeRequest
 from token_stats import AnalyzeTokens
 from users_sql import AnalyseUsersSql
 from coverage_journeys_transfers import AnalyzeCoverageJourneysTransfers
-from analyzer import Analyzer
